% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set for fish.}
\format{
A data set of 144 rows and 5 variables:
\describe{
  \item{LENGTH}{length of fish}
  \item{WEIGHT}{weight of fish}
  ...
}
}
\source{
Text book
}
\usage{
ddt
}
\description{
144 lines of measurements on fish.
}
\keyword{datasets}
