% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDMG data set for fire damage.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DAMAGE}{damage done by fire}
  \item{DISTANCE}{distance of fire spread}
  ...
}
}
\source{
Text book
}
\usage{
fire
}
\description{
Measurements of distance traveled by fire and
the damages done.
}
\keyword{datasets}
