% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{My Central Limit Theory Uniform Function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{how many iterations}

\item{a}{variable used to calculate mean, sd, variance, and the curve}

\item{b}{variable used to calculate mean, sd, variance, and the curve}
}
\value{
A histograph with mutliple thoeretical curves aligned to the sample distribution
}
\description{
My Central Limit Theory Uniform Function
}
