% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt function}
\usage{
myddt(df, x, y, cond, species, col)
}
\arguments{
\item{df}{ddt data set}

\item{x}{x variable is length of fish}

\item{y}{y variable is weight of fish}

\item{cond}{condition that can be fish bigger than a specified length}

\item{species}{variable used to select the species of fish used in the graph}

\item{col}{just a variable to pick the colors later}
}
\value{
A sophisticated plot, csv files, prints list of tables
}
\description{
myddt function
}
