% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{x variable}

\item{y}{y variable as a vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A sophisticated plot
}
\description{
Creates a sophisticated plot.
}
